Prefab change
fix(carrying): had to change the layout of the carrying object in order to make it work.
Note: the 'layout' is a bit different than the other interactables. The interact script is as a child object. Had to do that in order for the gravity to work on the object