Finished base code + base quest structure
I added the remaining base code for the quest system and added a base set up for the quests. These quests are sorted in multiple folders and have various different quest steps.

TO-DO:
- Figure out the input manager to make the submit button work as part of the QuestPoint.cs script to enable and disable the different quest steps and trigger points.
- Some quest step scrips need to be filled with code entirely.

All the way at the end: Set up all quests and rewards properly and add the last bits of code.