Level Design iterations
Added the pipe and the cable.
Created slopes to get up to cliff level 1 and 2.
Moved cliffs to create a valley and added the entrance to the cave area on the second cliff level.
Resized some of the bigger hot springs and replaced them with additional but smaller hot springs.