Preliminary Skybox
feat: two inward spheres for the skyboxes, each with a sky or night material. The top sphere (day) has a custom made shader using code that makes it fade into the night sphere. Since the standard Unity skybox is gray (but also casts shadows), I made the camera use a solid color (orange ) as the background to mimic a sunset.