Dialogue & Side quests
I've added a lot of dialogue because I had too much fun playing around with it. I also sorted all ink files in folders.

The narrative designers are free to change any existing dialogue that I've written besides the two toaster repair files (they can but I want to keep the essence of it). The last conversation with Chobo when starting the  "back to space" quest, is supposed to break the 4th wall, so please keep it similar to what I have (important for my research question) or make it even more in this direction.

Side quest base set up is done as well. Some scripts, questpoints, and placements have to be done.